feat(sdk): align endpoints and DTOs with backend api
